% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedded.R
\name{embedded}
\alias{embedded}
\alias{embedded,ts-method}
\title{embedding time series data}
\usage{
\S4method{embedded}{data.frame}(data, target, E = 3, tau = 0)
}
\arguments{
\item{data}{observation data.}

\item{target}{name of target variable.}

\item{E}{(optional) embedding dimensions.}

\item{tau}{(optional) step of time lags.}
}
\value{
A matrix
}
\description{
embedding time series data
}
\examples{
embedded(data.frame(t = 1:5),"t",3)

}
